tools

+++ tools/models.py
@@ -208,19 +208,19 @@ class Priem(models.Model):
             title_c=self.tool.title.split (' ',1)[0]
             alltools = Toolsonwarehouse.objects.filter(Q(title__icontains=title_c.lower()) | Q(title__icontains=title_c.upper()))
             count_all=0
-            for item in alltools:
-                
-                count_all+=item.count
-                if not item.pk == self.tool.pk:
-                    self.tool.text='Пред. место на '+dateformat.format(timezone.now(), 'd-m-Y')+' - '+str(item.workplace)+'\n'+str(self.tool.text)
-                    for v in Tools.objects.filter(tool=item):
-                        v.tool=self.tool
-                        v.save()
-                    for p in Priem.objects.filter(tool=item):
-                        if not p == self:
-                            p.tool=self.tool
-                            p.save()
-                    item.delete()
+            if alltools:
+                for item in alltools:
+                    count_all+=item.count
+                    if not item.pk == self.tool.pk:
+                        self.tool.text='Пред. место на '+dateformat.format(timezone.now(), 'd-m-Y')+' - '+str(item.workplace)+'\n'+str(self.tool.text)
+                        for v in Tools.objects.filter(tool=item):
+                            v.tool=self.tool
+                            v.save()
+                        for p in Priem.objects.filter(tool=item):
+                            if not p == self:
+                                p.tool=self.tool
+                                p.save()
+                        item.delete()
             self.tool.count=count_all + int(self.count or 0)
             if not self.tool.workplace==self.place:
                 self.tool.text='Пред. место на '+dateformat.format(timezone.now(), 'd-m-Y')+' - '+str(self.tool.workplace)+'\n'+str(self.tool.text)
@@ -233,16 +233,17 @@ class Priem(models.Model):
             title_c=self.tool.title.split (' ',1)[0]
             alltools = Toolsonwarehouse.objects.filter(Q(title__icontains=title_c.lower()) | Q(title__icontains=title_c.upper()))
             count_all=0
-            for item in alltools:
-                count_all+=item.count
-                if not item.pk == self.tool.pk:
-                    self.tool.text='Пред. место на '+dateformat.format(timezone.now(), 'd-m-Y')+' - '+str(item.workplace)+'\n'+str(self.tool.text)
-                    for v in Tools.objects.filter(tool=item):
-                        v.tool=self.tool
-                        v.save()
-                    
+            if alltools:
+                for item in alltools:
+                    count_all+=int(item.count or 0)
+                    if not item.pk == self.tool.pk:
+                        self.tool.text='Пред. место на '+dateformat.format(timezone.now(), 'd-m-Y')+' - '+str(item.workplace)+'\n'+str(self.tool.text)
+                        for v in Tools.objects.filter(tool=item):
+                            v.tool=self.tool
+                            v.save()
                         
-                    item.delete()
+                            
+                        item.delete()
             pr=Priem.objects.filter(pk=self.id).first()
             if pr:
                 previous_count = pr.count
